Delete 'promised_achvmts', update 'ed_type' fields

+++ perfomance_tracker/groups/models.py
@@ -36,7 +36,7 @@ class Group(models.Model):
         max_length=5,
         verbose_name='Подгруппа'
     )
-    education_type = models.IntegerField(
+    education_type = models.PositiveSmallIntegerField(
         choices=EDUCATION_TYPE_CHOICES,
         verbose_name='Тип обучения'
     )
